Added go home function to Navbar's boogie woogie
